# Fix InvalidRequestException in aws_emr_instance_group table. Fixes #1222