# Integration test logs
<details>
  <summary>Logs</summary>
  
```
SETUP: tests/aws_emr_instance_group []

PRETEST: tests/aws_emr_instance_group

TEST: tests/aws_emr_instance_group
Running terraform
data.aws_partition.current: Reading...
data.aws_region.alternate: Reading...
data.aws_partition.current: Read complete after 0s [id=aws]
data.aws_region.alternate: Read complete after 0s [id=us-east-2]
data.aws_region.primary: Reading...
data.aws_caller_identity.current: Reading...
data.aws_region.primary: Read complete after 0s [id=us-east-1]
data.aws_caller_identity.current: Read complete after 2s [id=533793682495]
data.null_data_source.resource: Reading...
data.null_data_source.resource: Read complete after 0s [id=static]

Terraform used the selected providers to generate the following execution
plan. Resource actions are indicated with the following symbols:
  + create

Terraform will perform the following actions:

  # aws_emr_cluster.named_test_resource will be created
  + resource "aws_emr_cluster" "named_test_resource" {
      + applications                      = [
          + "Spark",
        ]
      + arn                               = (known after apply)
      + cluster_state                     = (known after apply)
      + id                                = (known after apply)
      + keep_job_flow_alive_when_no_steps = (known after apply)
      + master_public_dns                 = (known after apply)
      + name                              = "turbottest16186"
      + release_label                     = "emr-4.9.5"
      + scale_down_behavior               = (known after apply)
      + service_role                      = (known after apply)
      + step                              = (known after apply)
      + step_concurrency_level            = 1
      + tags_all                          = (known after apply)
      + termination_protection            = (known after apply)
      + visible_to_all_users              = true

      + core_instance_fleet {
          + id                             = (known after apply)
          + name                           = (known after apply)
          + provisioned_on_demand_capacity = (known after apply)
          + provisioned_spot_capacity      = (known after apply)
          + target_on_demand_capacity      = (known after apply)
          + target_spot_capacity           = (known after apply)

          + instance_type_configs {
              + bid_price                                  = (known after apply)
              + bid_price_as_percentage_of_on_demand_price = (known after apply)
              + instance_type                              = (known after apply)
              + weighted_capacity                          = (known after apply)

              + configurations {
                  + classification = (known after apply)
                  + properties     = (known after apply)
                }

              + ebs_config {
                  + iops                 = (known after apply)
                  + size                 = (known after apply)
                  + type                 = (known after apply)
                  + volumes_per_instance = (known after apply)
                }
            }

          + launch_specifications {
              + on_demand_specification {
                  + allocation_strategy = (known after apply)
                }

              + spot_specification {
                  + allocation_strategy      = (known after apply)
                  + block_duration_minutes   = (known after apply)
                  + timeout_action           = (known after apply)
                  + timeout_duration_minutes = (known after apply)
                }
            }
        }

      + core_instance_group {
          + autoscaling_policy = (known after apply)
          + bid_price          = (known after apply)
          + id                 = (known after apply)
          + instance_count     = (known after apply)
          + instance_type      = (known after apply)
          + name               = (known after apply)

          + ebs_config {
              + iops                 = (known after apply)
              + size                 = (known after apply)
              + throughput           = (known after apply)
              + type                 = (known after apply)
              + volumes_per_instance = (known after apply)
            }
        }

      + ec2_attributes {
          + emr_managed_master_security_group = (known after apply)
          + emr_managed_slave_security_group  = (known after apply)
          + instance_profile                  = (known after apply)
          + key_name                          = "deployer-key"
          + service_access_security_group     = (known after apply)
          + subnet_id                         = (known after apply)
          + subnet_ids                        = (known after apply)
        }

      + master_instance_fleet {
          + id                             = (known after apply)
          + name                           = (known after apply)
          + provisioned_on_demand_capacity = (known after apply)
          + provisioned_spot_capacity      = (known after apply)
          + target_on_demand_capacity      = (known after apply)
          + target_spot_capacity           = (known after apply)

          + instance_type_configs {
              + bid_price                                  = (known after apply)
              + bid_price_as_percentage_of_on_demand_price = (known after apply)
              + instance_type                              = (known after apply)
              + weighted_capacity                          = (known after apply)

              + configurations {
                  + classification = (known after apply)
                  + properties     = (known after apply)
                }

              + ebs_config {
                  + iops                 = (known after apply)
                  + size                 = (known after apply)
                  + type                 = (known after apply)
                  + volumes_per_instance = (known after apply)
                }
            }

          + launch_specifications {
              + on_demand_specification {
                  + allocation_strategy = (known after apply)
                }

              + spot_specification {
                  + allocation_strategy      = (known after apply)
                  + block_duration_minutes   = (known after apply)
                  + timeout_action           = (known after apply)
                  + timeout_duration_minutes = (known after apply)
                }
            }
        }

      + master_instance_group {
          + id             = (known after apply)
          + instance_count = 1
          + instance_type  = "m4.large"

          + ebs_config {
              + iops                 = (known after apply)
              + size                 = (known after apply)
              + throughput           = (known after apply)
              + type                 = (known after apply)
              + volumes_per_instance = (known after apply)
            }
        }
    }

  # aws_iam_instance_profile.emr_profile will be created
  + resource "aws_iam_instance_profile" "emr_profile" {
      + arn         = (known after apply)
      + create_date = (known after apply)
      + id          = (known after apply)
      + name        = "var.resource_name"
      + path        = "/"
      + role        = "turbottest16186_2"
      + tags_all    = (known after apply)
      + unique_id   = (known after apply)
    }

  # aws_iam_role.iam_emr_profile_role will be created
  + resource "aws_iam_role" "iam_emr_profile_role" {
      + arn                   = (known after apply)
      + assume_role_policy    = jsonencode(
            {
              + Statement = [
                  + {
                      + Action    = "sts:AssumeRole"
                      + Effect    = "Allow"
                      + Principal = {
                          + Service = "ec2.amazonaws.com"
                        }
                      + Sid       = "test"
                    },
                ]
              + Version   = "2008-10-17"
            }
        )
      + create_date           = (known after apply)
      + force_detach_policies = false
      + id                    = (known after apply)
      + managed_policy_arns   = (known after apply)
      + max_session_duration  = 3600
      + name                  = "turbottest16186_2"
      + name_prefix           = (known after apply)
      + path                  = "/"
      + tags_all              = (known after apply)
      + unique_id             = (known after apply)

      + inline_policy {
          + name   = (known after apply)
          + policy = (known after apply)
        }
    }

  # aws_iam_role.iam_emr_service_role will be created
  + resource "aws_iam_role" "iam_emr_service_role" {
      + arn                   = (known after apply)
      + assume_role_policy    = jsonencode(
            {
              + Statement = [
                  + {
                      + Action    = "sts:AssumeRole"
                      + Effect    = "Allow"
                      + Principal = {
                          + Service = "elasticmapreduce.amazonaws.com"
                        }
                      + Sid       = ""
                    },
                ]
              + Version   = "2008-10-17"
            }
        )
      + create_date           = (known after apply)
      + force_detach_policies = false
      + id                    = (known after apply)
      + managed_policy_arns   = (known after apply)
      + max_session_duration  = 3600
      + name                  = "turbottest16186_1"
      + name_prefix           = (known after apply)
      + path                  = "/"
      + tags_all              = (known after apply)
      + unique_id             = (known after apply)

      + inline_policy {
          + name   = (known after apply)
          + policy = (known after apply)
        }
    }

  # aws_iam_role_policy.iam_emr_profile_policy will be created
  + resource "aws_iam_role_policy" "iam_emr_profile_policy" {
      + id     = (known after apply)
      + name   = "turbottest16186_2"
      + policy = jsonencode(
            {
              + Statement = [
                  + {
                      + Action   = [
                          + "cloudwatch:*",
                          + "ec2:Describe*",
                          + "elasticmapreduce:Describe*",
                          + "elasticmapreduce:ListBootstrapActions",
                          + "elasticmapreduce:ListClusters",
                          + "elasticmapreduce:ListInstanceGroups",
                          + "elasticmapreduce:ListInstances",
                          + "elasticmapreduce:ListSteps",
                          + "s3:*",
                          + "sdb:*",
                          + "sns:*",
                          + "sqs:*",
                        ]
                      + Effect   = "Allow"
                      + Resource = "*"
                    },
                ]
              + Version   = "2012-10-17"
            }
        )
      + role   = (known after apply)
    }

  # aws_iam_role_policy.iam_emr_service_policy will be created
  + resource "aws_iam_role_policy" "iam_emr_service_policy" {
      + id     = (known after apply)
      + name   = "turbottest16186_1"
      + policy = jsonencode(
            {
              + Statement = [
                  + {
                      + Action   = [
                          + "ec2:AuthorizeSecurityGroupEgress",
                          + "ec2:AuthorizeSecurityGroupIngress",
                          + "ec2:CancelSpotInstanceRequests",
                          + "ec2:CreateNetworkInterface",
                          + "ec2:CreateSecurityGroup",
                          + "ec2:CreateTags",
                          + "ec2:DeleteNetworkInterface",
                          + "ec2:DeleteSecurityGroup",
                          + "ec2:DeleteTags",
                          + "ec2:DescribeAvailabilityZones",
                          + "ec2:DescribeAccountAttributes",
                          + "ec2:DescribeDhcpOptions",
                          + "ec2:DescribeInstanceStatus",
                          + "ec2:DescribeInstances",
                          + "ec2:DescribeKeyPairs",
                          + "ec2:DescribeNetworkAcls",
                          + "ec2:DescribeNetworkInterfaces",
                          + "ec2:DescribePrefixLists",
                          + "ec2:DescribeRouteTables",
                          + "ec2:DescribeSecurityGroups",
                          + "ec2:DescribeSpotInstanceRequests",
                          + "ec2:DescribeSpotPriceHistory",
                          + "ec2:DescribeSubnets",
                          + "ec2:DescribeVpcAttribute",
                          + "ec2:DescribeVpcEndpoints",
                          + "ec2:DescribeVpcEndpointServices",
                          + "ec2:DescribeVpcs",
                          + "ec2:DetachNetworkInterface",
                          + "ec2:ModifyImageAttribute",
                          + "ec2:ModifyInstanceAttribute",
                          + "ec2:RequestSpotInstances",
                          + "ec2:RevokeSecurityGroupEgress",
                          + "ec2:RunInstances",
                          + "ec2:TerminateInstances",
                          + "ec2:DeleteVolume",
                          + "ec2:DescribeVolumeStatus",
                          + "ec2:DescribeVolumes",
                          + "ec2:DetachVolume",
                          + "iam:GetRole",
                          + "iam:GetRolePolicy",
                          + "iam:ListInstanceProfiles",
                          + "iam:ListRolePolicies",
                          + "iam:PassRole",
                          + "s3:CreateBucket",
                          + "s3:Get*",
                          + "s3:List*",
                          + "sdb:BatchPutAttributes",
                          + "sdb:Select",
                          + "sqs:CreateQueue",
                          + "sqs:Delete*",
                          + "sqs:GetQueue*",
                          + "sqs:PurgeQueue",
                          + "sqs:ReceiveMessage",
                        ]
                      + Effect   = "Allow"
                      + Resource = "*"
                    },
                ]
              + Version   = "2012-10-17"
            }
        )
      + role   = (known after apply)
    }

  # aws_internet_gateway.gw will be created
  + resource "aws_internet_gateway" "gw" {
      + arn      = (known after apply)
      + id       = (known after apply)
      + owner_id = (known after apply)
      + tags_all = (known after apply)
      + vpc_id   = (known after apply)
    }

  # aws_key_pair.deployer will be created
  + resource "aws_key_pair" "deployer" {
      + arn             = (known after apply)
      + fingerprint     = (known after apply)
      + id              = (known after apply)
      + key_name        = "deployer-key"
      + key_name_prefix = (known after apply)
      + key_pair_id     = (known after apply)
      + key_type        = (known after apply)
      + public_key      = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC6B60dQX19dOO0wmNnPf27NJBXQhuVAsFcshY7YqTmNA1dfl0lXvJFSF0nVTvaxtE2854gmvygm/yafopG33zJeH+k2ZrblivkEuH0NNajKMp2mBumkk8RiTVvU1ET4m6Z1jZaK0dTSeV2k6gUk7Wmvo4+6RIu+wdAtwGGGcSq4HJ4M0G1CGHruBywMdOs5CklqRn7BRMh2aexowDachlabUKRFwTf7OCjDcoCWHo+o4kHV7NDEdhFzrE2fxXttt054kd5CMZOEOB9p01pAqSPrxFv3S/BpiXuhmjwsUsIdfHIgFrFhNY/M4+KhTa3pI69NqsigReotTEzk2QwHyHSjsRwfzeJNRgI8iqGIdYr3+t6zCKswyzgV0hqalZhXtpVUEBbumJAKWy+VzTnOXUQ5zEVvabTxi3dsW9zMGofjwhiC75x97RE7U/ZEanRM+0Avr8e3c3ob4NLf9JOwbvB/jYvS6j/nuwJ5H6igT3Oj0oDiknh2WtOgH/BKKaCq5Nyl15Df65i0PNgKRcLy+x0fubYUtsCOMPpQTfztSYGjbvAhAOrH2LJF0i2dTD3PcVw6R6uDpJSw9QkzdD+2ofxnqjGPUeKR3mraGOXt2hZe9F9bvvdZScDfdAoyBUYTO2/sxsQFobRtdiEk2up+mNIx9QXMXO2dOgtNUYZKc1DdQ== clara@turbot.com"
      + tags_all        = (known after apply)
    }

  # aws_main_route_table_association.test_association will be created
  + resource "aws_main_route_table_association" "test_association" {
      + id                      = (known after apply)
      + original_route_table_id = (known after apply)
      + route_table_id          = (known after apply)
      + vpc_id                  = (known after apply)
    }

  # aws_route_table.test_route_table will be created
  + resource "aws_route_table" "test_route_table" {
      + arn              = (known after apply)
      + id               = (known after apply)
      + owner_id         = (known after apply)
      + propagating_vgws = (known after apply)
      + route            = [
          + {
              + carrier_gateway_id         = ""
              + cidr_block                 = "0.0.0.0/0"
              + core_network_arn           = ""
              + destination_prefix_list_id = ""
              + egress_only_gateway_id     = ""
              + gateway_id                 = (known after apply)
              + instance_id                = ""
              + ipv6_cidr_block            = ""
              + local_gateway_id           = ""
              + nat_gateway_id             = ""
              + network_interface_id       = ""
              + transit_gateway_id         = ""
              + vpc_endpoint_id            = ""
              + vpc_peering_connection_id  = ""
            },
        ]
      + tags_all         = (known after apply)
      + vpc_id           = (known after apply)
    }

  # aws_security_group.emr_master will be created
  + resource "aws_security_group" "emr_master" {
      + arn                    = (known after apply)
      + description            = "Managed by Terraform"
      + egress                 = [
          + {
              + cidr_blocks      = [
                  + "0.0.0.0/0",
                ]
              + description      = ""
              + from_port        = 0
              + ipv6_cidr_blocks = []
              + prefix_list_ids  = []
              + protocol         = "-1"
              + security_groups  = []
              + self             = false
              + to_port          = 0
            },
        ]
      + id                     = (known after apply)
      + ingress                = [
          + {
              + cidr_blocks      = []
              + description      = ""
              + from_port        = 443
              + ipv6_cidr_blocks = []
              + prefix_list_ids  = []
              + protocol         = "tcp"
              + security_groups  = []
              + self             = false
              + to_port          = 443
            },
        ]
      + name                   = (known after apply)
      + name_prefix            = (known after apply)
      + owner_id               = (known after apply)
      + revoke_rules_on_delete = true
      + tags_all               = (known after apply)
      + vpc_id                 = (known after apply)
    }

  # aws_subnet.my_subnet will be created
  + resource "aws_subnet" "my_subnet" {
      + arn                                            = (known after apply)
      + assign_ipv6_address_on_creation                = false
…